+++ CHANGELOG.md
@@ -11,7 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Backend: Update Spring Boot 4.0.0, Spring Modulith 2.0.0, BouncyCastle 1.82,  Hibernate 7.1.9.Final, ehcache 3.11.1, H2 2.4.240, Gradle Plugins: CycloneDX 3.0.2, Spotless 8.1.0, Update Plugin 0.53.0, Google Java Format 1.32.0
 * Backend: Added to documentation the new JDK 25 (also backported to JDK 21/17) JEP519 Option for Compact Object Headers leading to significant performance improvements and memory reduction
 * Frontend: Update [Angular 21](https://blog.angular.dev/announcing-angular-v21-57946c34f14b)
+* Frontend: Move to vitest from karma
 * Frontend: Remove zone.js
+* Frontend: Fix view refresh for data loaded during component initialisation
 
 ## [0.0.10] - 2025-08-18
 * Backend: Update to [Gradle 9.0.0](https://docs.gradle.org/9.0.0/release-notes.html). Also enabling the new [configuration cache feature](https://docs.gradle.org/current/userguide/configuration_cache.html)